Code cleaning with Adytia patch

+++ choco/src/main/java/org/btrplace/scheduler/choco/extensions/RoundedUpDivision.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 University Nice Sophia Antipolis
+ * Copyright (c) 2017 University Nice Sophia Antipolis
  *
  * This file is part of btrplace.
  * This library is free software; you can redistribute it and/or
@@ -77,13 +77,13 @@ public class RoundedUpDivision extends Constraint {
         @Override
         public ESat isEntailed() {
             if (vars[0].getDomainSize() == 1 && vars[1].getDomainSize() == 1) {
-                return ESat.eval(vars[0].getValue() == (int) Math.ceil((double) vars[1].getValue() / divider));
+                return ESat.eval(vars[0].getValue() == (int) Math.ceil(vars[1].getValue() / divider));
             }
             return ESat.UNDEFINED;
         }
 
         private int div(int b) {
-            return (int) Math.ceil((double) b / divider);
+            return (int) Math.ceil(b / divider);
         }
 
         private int multLB(int a) {
